requirements.txt
# Apache Airflow 
apache-airflow==3.1.5 --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-3.1.5/constraints-3.12.txt"

# Providers
apache-airflow-providers-http
apache-airflow-providers-amazon
apache-airflow-providers-amazon[s3fs]

# Data processing
pandas>=1.5,<3.0
pyarrow>=12.0

# Analytics
duckdb>=0.9

# Utilities
requests>=2.28
pendulum>=2.1